+++ Cargo-minimal.lock
@@ -92,12 +92,16 @@ name = "bitcoin-crypto"
 version = "0.0.0"
 dependencies = [
  "arbitrary",
+ "base58ck",
  "bitcoin-internals",
  "bitcoin-io",
+ "bitcoin-network-kind",
+ "bitcoin_hashes",
  "hex-conservative 0.3.2",
  "hex-conservative 1.0.0",
  "secp256k1",
  "serde",
+ "serde_test",
 ]
 
 [[package]]
