+++ crypto/Cargo.toml
@@ -15,22 +15,27 @@ exclude = ["tests", "contrib"]
 
 [features]
 default = ["std"]
-std = ["alloc", "hex-stable/std", "hex-unstable/std", "internals/std", "io/std", "secp256k1/std", "serde?/std"]
-alloc = ["hex-stable/alloc", "hex-unstable/alloc", "internals/alloc", "io/alloc", "secp256k1/alloc", "serde?/alloc"]
-serde = ["dep:serde", "internals/serde", "secp256k1/serde"]
+std = ["alloc", "base58/std", "hashes/std", "hex-stable/std", "hex-unstable/std", "internals/std", "io/std", "network/std", "secp256k1/std", "serde?/std"]
+rand = ["secp256k1/rand"]
+alloc = ["base58/alloc", "hashes/alloc", "hex-stable/alloc", "hex-unstable/alloc", "internals/alloc", "io/alloc", "network/alloc", "secp256k1/alloc", "serde?/alloc"]
+serde = ["dep:serde", "hashes/serde", "internals/serde", "secp256k1/serde"]
 arbitrary = ["dep:arbitrary", "secp256k1/arbitrary"]
 
 [dependencies]
+base58 = { package = "base58ck", path = "../base58", version = "0.4.0", default-features = false }
+hashes = { package = "bitcoin_hashes", path = "../hashes", version = "0.20.0", default-features = false, features = ["hex"] }
 hex-unstable = { package = "hex-conservative", version = "0.3.2", default-features = false }
 hex-stable = { package = "hex-conservative", version = "1.0.0", default-features = false }
 internals = { package = "bitcoin-internals", path = "../internals", version = "0.5.0", features = ["hex"] }
 io = { package = "bitcoin-io", path = "../io", version = "0.5.0", default-features = false, features = ["hashes"] }
+network = { package = "bitcoin-network-kind", path = "../network", version = "0.1.0", default-features = false }
 secp256k1 = { version = "0.32.0-beta.2", default-features = false }
 
 arbitrary = { version = "1.4.1", optional = true }
 serde = { version = "1.0.195", default-features = false, features = ["derive"], optional = true }
 
 [dev-dependencies]
+serde_test = "1.0.19"
 
 [package.metadata.docs.rs]
 all-features = true
